+++ test/test_helper.exs
@@ -1,1 +1,3 @@
+ExUnit.configure(exclude: [:hardware])
+Application.ensure_all_started(:telemetry)
 ExUnit.start()